requirements.txt
aiofiles
colorama
gpytranslate
flask
chardet
discord.py-self
aiodns
